+++ tests/Unit/Public/New-ScriptLog.Tests.ps1
@@ -1,15 +1,25 @@
 Describe 'New-ScriptLog' {
-    It 'Given no parameters, it should return a ScriptLog object with default properties' {
+    It 'should return a ScriptLog object with default properties when no parameters are given' {
         $ScriptLog = New-ScriptLog
         $ScriptLog.GetType().Name | Should -Be 'ScriptLog'
+        $ScriptLog.LogType | Should -Be 'CMTrace'
+        $ScriptLog.MessagesOnConsole | Should -Be @('Error', 'Warning')
     }
-    It 'Given all parameters set to non-default value, it should return a ScriptLog object that matches the parameters set' {
+    It 'should return a ScriptLog object that matches the given parameters, when all parameters are set to non-default value' {
         $ScriptLog = New-ScriptLog -Path $PSScriptRoot -BaseName 'CustomLog' -LogType Memory -MessagesOnConsole @('Information', 'Verbose')
         $ScriptLog.GetType().Name | Should -Be 'ScriptLog'
+        $ScriptLog.LogType | Should -Be 'Memory'
+        $ScriptLog.MessagesOnConsole | Should -Be @('Information', 'Verbose')
     }
-    It 'Given the AppendDateTime parameter, it should return a ScriptLog object where the outfile name have a datetime suffixed' {
+    It 'should return a ScriptLog object where the outfile name have a datetime suffixed, if AppendDateTime is set to true' {
         $ScriptLog = New-ScriptLog -AppendDateTime $true
         $ScriptLog.GetType().Name | Should -Be 'ScriptLog'
+        $ScriptLog.FilePath -match 'ScriptLog-\d{14}\.log$' | Should -Be $true
     }
-    #TODO: Improve quality of tests and add additional tests to New-ScriptLog
+    It 'should throw an error if two ScriptLog objects are using the same filepath' {
+        { 
+            $ScriptLog1 = New-ScriptLog
+            $ScriptLog2 = New-ScriptLog 
+        } | Should -Throw
+    }
 }