+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Fixed
+
+- It is no longer possible to have two ScrtipLog objects using the same file.
+
 ### Changed
+
+- Added better testing to New-ScriptLog.
 
 ## [0.3.1] - 2022-07-14